+++ src/java/com/github/jsdossier/LinkFactory.java
@@ -295,8 +295,10 @@ final class LinkFactory {
       return typeRef;
     }
 
-    if (checkPrototype && (type.getType().isConstructor() || type.getType().isInterface())) {
-      ObjectType instanceType = ((FunctionType) type.getType()).getInstanceType();
+    if (checkPrototype && (type.getType().isInstanceType() || type.getType().isConstructor() || type.getType().isInterface())) {
+      ObjectType instanceType = type.getType().isInstanceType()
+          ? type.getType().toObjectType()
+          : ((FunctionType) type.getType()).getInstanceType();
       if (instanceType.getPropertyNames().contains(property)) {
         return typeRef
             .toBuilder()